+++ ui/packages/catalog-ui-search/src/main/webapp/extension-points/extension-points.tsx
@@ -26,6 +26,7 @@ import searchInteractions, {
 import { tableExport, Props as TableExportProps } from './table-export'
 import multiSelectActions from './multi-select-actions'
 import sharingAliases from './sharing-aliases'
+import getSuggestions from './suggestions'
 
 export type ExtensionPointsType = {
   routes: {}
@@ -39,6 +40,7 @@ export type ExtensionPointsType = {
   tableExport: SFC<TableExportProps>
   multiSelectActions: any[]
   sharingAliases: any[]
+  getSuggestions: (attribute: any, suggester: any) => Promise<any>
 }
 
 const ExtensionPoints: ExtensionPointsType = {
@@ -53,6 +55,7 @@ const ExtensionPoints: ExtensionPointsType = {
   tableExport,
   multiSelectActions,
   sharingAliases,
+  getSuggestions,
 }
 
 export default ExtensionPoints
